requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
streamlit-extras
sentence-transformers
tiktoken
python-dotenv
langchain-google-genai
langchain-community
google-palm
SpeechRecognition
moviepy

# langchain==0.0.184
# PyPDF2==3.0.1
# python-dotenv==1.0.0
# streamlit==1.18.1
# openai==0.27.6
# faiss-cpu==1.7.4
# altair==4
# tiktoken==0.4.0